feat: add gitignore for autoformat code

+++ projects/iv-week/functions/pure-double.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.4;
 
 contract Contract {
-    function double(uint x) external pure returns(uint sum) {
-    sum = 2 * x;
-}
+    function double(uint256 x) external pure returns (uint256 sum) {
+        sum = 2 * x;
+    }
 }
